requirements.txt
# 以下为训练和预处理专用
SimpleITK>=2.0.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.3.0
tqdm>=4.62.0
torchmetrics>=0.11.0

# 用于运行 Web 界面（推理），不包含训练和预处理相关库
torch>=1.9.0
torchvision>=0.14.0
ultralytics>=8.0.0
streamlit>=1.25.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=9.0.0
reportlab>=3.6.0
pyyaml>=6.0 